% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typo.R
\name{contingency_table}
\alias{contingency_table}
\title{Contingency Tables}
\usage{
contingency_table(ids, names = TRUE, absolute = TRUE, data = WALS)
}
\arguments{
\item{ids}{Vector of feature IDs}

\item{names}{Boolean; should the output use prose value descriptions? If \code{FALSE}, value IDs are used instead.}

\item{absolute}{Boolean; should the output be in the form of absolute frequencies? If \code{FALSE}, relative frequencies are outputted.}

\item{data}{Optionally a data frame, e.g. a subset of \code{\link{WALS}}}
}
\value{
An \emph{n}-by-\emph{n} contingency table, where \emph{n} is the length of \code{ids}
}
\description{
Produce a contingency table from the values of two or more WALS features.
}
\details{
This function calls \code{\link{intersect_features}} internally to establish the database intersection based on which a contingency table may be created.
}
\examples{
# 2x2 contingency table (absolute frequencies) of the first two features:
tab <- contingency_table(c("1A", "2A"))

# Run a chi-square test on the above:
chisq.test(tab)

# 3x3 contingency table of the first three features, showing relative
# frequencies and value IDs instead of value descriptions:
contingency_table(c("1A", "2A", "3A"), names=FALSE, absolute=FALSE)
}
