% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WALS.R
\name{country_codes}
\alias{country_codes}
\title{Country Codes}
\usage{
country_codes(id)
}
\arguments{
\item{id}{Language's WALS ID, or a vector of IDs}
}
\value{
A list of character vectors
}
\description{
Return the country codes associated with a language
}
\examples{
country_codes("fre")
unlist(country_codes("fre"))
country_codes(c("fre", "ger"))
}
