% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WALS.R
\name{feature_metadata}
\alias{feature_metadata}
\title{WALS Feature Metadata}
\usage{
feature_metadata(id)
}
\arguments{
\item{id}{Feature's WALS ID, or a vector of IDs}
}
\value{
A dataframe with the following columns:
\describe{
\item{\code{feature_ID}}{Feature's WALS ID}
\item{\code{feature}}{Feature's name}
\item{\code{value_ID}}{Value ID}
\item{\code{value}}{Prose description of value}
}
}
\description{
Returns information about a WALS feature. Use this to find the possible
values the feature can take.
}
\examples{
feature_metadata("13A")
feature_metadata("13A")$value_ID
feature_metadata(c("13A", "9A", "83A"))
}
