% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WALS.R
\name{get_feature}
\alias{get_feature}
\title{Get Feature}
\usage{
get_feature(id, data = WALS)
}
\arguments{
\item{id}{Feature's WALS ID}

\item{data}{Optionally, a data frame (such as a subset of \code{\link{WALS}})}
}
\value{
A data frame
}
\description{
Return the representation of a feature in WALS.
}
\details{
Returns the relevant portion of the \code{\link{WALS}} data frame.
Feature values and value IDs are returned as factors.
}
