% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WALS.R
\name{get_language}
\alias{get_language}
\title{Get WALS Language}
\usage{
get_language(id)
}
\arguments{
\item{id}{Language's WALS ID.}
}
\value{
The portion of the \code{\link{WALS}} dataframe pertaining to this language.
}
\description{
Returns a dataframe containing the WALS representation of a language.
}
