% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WALS.R
\name{intersect_features}
\alias{intersect_features}
\title{Intersect Features}
\usage{
intersect_features(ids, data = WALS)
}
\arguments{
\item{ids}{Vector of IDs of features to intersect}

\item{data}{Optionally, a data frame (such as a subset of \code{\link{WALS}})}
}
\value{
A portion of the \code{\link{WALS}} data frame.
}
\description{
Return a subset of WALS consisting of only those languages for which data is attested in each of a given set of features; in effect, produce a set-theoretic intersection of the features.
}
\examples{
# Intersection of features 23A and 32A:
intersect_features(c("23A", "32A"))

# Intersection of features 1A, 9A and 13A:
intersect_features(c("1A", "9A", "13A"))

# Intersection of features 1A ... 20A:
df <- intersect_features(paste0(1:20, "A"))

# Find out which languages belong to that intersection:
unique(df$language_ID)

# And get their metadata:
language_metadata(unique(df$language_ID))

# Prove there is no language in WALS for which all features are attested:
df <- intersect_features(unique(WALS$feature_ID))
nrow(df) # == 0
}
