% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WALS.R
\name{language_metadata}
\alias{language_metadata}
\title{WALS Language Metadata}
\usage{
language_metadata(id)
}
\arguments{
\item{id}{Language's WALS ID, or a vector of IDs}
}
\value{
A dataframe with the following columns:
\describe{
\item{\code{language_ID}}{Language's WALS ID}
\item{\code{iso_code}}{Language's ISO code}
\item{\code{glottocode}}{Language's Glottocode}
\item{\code{language}}{Language's name}
\item{\code{latitude}}{Language's latitude}
\item{\code{longitude}}{Language's longitude}
\item{\code{genus}}{Language's genus}
\item{\code{family}}{Language's family}
\item{\code{macroarea}}{Language's macroarea}
\item{\code{countrycodes}}{Country codes associated with the language}
\item{\code{in_WALS_100}}{Boolean flag: \code{TRUE} if language is included in the WALS 100 sample}
\item{\code{in_WALS_200}}{Boolean flag: \code{TRUE} if language is included in the WALS 200 sample}
}
}
\description{
Returns information about a WALS language.
}
\examples{
language_metadata("fin")
language_metadata("fin")$family
language_metadata(c("fin", "swe"))
}
