% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{nearest_geo_neighbours}
\alias{nearest_geo_neighbours}
\title{Nearest Geographical Neighbours}
\usage{
nearest_geo_neighbours(id, k = 1, languages = WALS_languages)
}
\arguments{
\item{id}{Language's WALS ID}

\item{k}{How many neighbours to return}

\item{languages}{Set of languages to consider; by default all}
}
\description{
Return the \emph{k} nearest geographical neighbours of a language.
}
\details{
WALS defines the geographical representation of a language as a point
in latitude-longitude space.
Pairwise geographical distance between languages is defined as the 
great-circle distance (see \code{\link{great_circle_distance}}).
By default (\code{languages = WALS_languages}), all languages in the WALS
atlas will be considered in the computation of the k nearest neighbours.
By supplying a non-default \code{languages} argument, it is possible to
override this behaviour, e.g. to supply only a subset of the languages.
}
