% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{nearest_geo_neighbours_discoidal}
\alias{nearest_geo_neighbours_discoidal}
\title{Nearest Geographical Neighbours (Discoidal)}
\usage{
nearest_geo_neighbours_discoidal(id, radius = 1000,
  languages = WALS_languages)
}
\arguments{
\item{id}{Language's WALS ID}

\item{radius}{Radius of neighbourhood disc in kilometres}

\item{languages}{Set of languages to consider; by default all}
}
\description{
Return the neighbours of a language residing within a disc of a specified
radius from that language.
}
\details{
WALS defines the geographical representation of a language as a point
in latitude-longitude space.
Pairwise geographical distance between languages is defined as the 
great-circle distance (see \code{\link{great_circle_distance}}).
By default (\code{languages = WALS_languages}), all languages in the WALS
atlas will be considered in the computation of the k nearest neighbours.
By supplying a non-default \code{languages} argument, it is possible to
override this behaviour, e.g. to supply only a subset of the languages.
}
