% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WALS.R
\name{sample_WALS_100}
\alias{sample_WALS_100}
\title{WALS 100-Language Sample}
\usage{
sample_WALS_100()
}
\value{
A subset of the \code{\link{WALS}} data frame.
}
\description{
Return the WALS 100-language sample.
}
\details{
Returns the portion of \code{\link{WALS}} where \code{in_WALS_100} is \code{TRUE}.
}
