% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WALS.R
\name{sample_WALS_200}
\alias{sample_WALS_200}
\title{WALS 200-Language Sample}
\usage{
sample_WALS_200()
}
\value{
A subset of the \code{\link{WALS}} data frame.
}
\description{
Return the WALS 200-language sample.
}
\details{
Returns the portion of \code{\link{WALS}} where \code{in_WALS_200} is \code{TRUE}. Note that the WALS 200-languages sample contains \emph{202} languages.
}
