% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{WALS_100}
\alias{WALS_100}
\title{World Atlas of Language Structures (100-Language Sample)}
\format{See \code{\link{WALS}}.}
\source{
\url{http://wals.info}
}
\description{
The WALS 100-language sample: the portion of \code{\link{WALS}} pertaining to languages belonging to this sample.
}
