% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{WALS_nometa}
\alias{WALS_nometa}
\title{World Atlas of Language Structures (Without Metadata)}
\format{A data frame with 76465 rows and 5 variables:
\describe{
\item{\code{language_ID}}{}
\item{\code{feature_ID}}{}
\item{\code{value_ID}}{}
\item{\code{latitude}}{}
\item{\code{longitude}}{}
}}
\source{
\url{http://wals.info}
}
\description{
The WALS dataset without metadata. It is recommended to use this instead of \code{\link{WALS}} for any computationally intensive operations.
}
