% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WALS.R
\name{get_feature}
\alias{get_feature}
\title{Get WALS Feature}
\usage{
get_feature(id)
}
\arguments{
\item{id}{Feature's WALS ID.}
}
\value{
The portion of the \code{\link{WALS}} dataframe pertaining to this feature.
}
\description{
Returns a dataframe containing the WALS coverage of a feature.
}
