% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{nearest_geo_neighbours}
\alias{nearest_geo_neighbours}
\title{Nearest Geographical Neighbours}
\usage{
nearest_geo_neighbours(id, k, data = WALS)
}
\arguments{
\item{id}{Language's WALS ID}

\item{k}{How many neighbours to return}

\item{data}{Dataset to consider; by default the entire \code{\link{WALS}}}
}
\description{
Return the \emph{k} nearest geographical neighbours of a language.
}
\details{
Computed using \code{\link{great_circle_distance}}.
By default (\code{data = WALS}), all languages in the WALS
atlas will be considered in the computation of the nearest neighbours.
By supplying a non-default \code{data} argument, it is possible to
override this behaviour, e.g. to supply only a subset of the languages.

If the language does not have as many as \emph{k} neighbours (e.g. because
of a strict filtering by \code{data}), the neighbours that \emph{are}
found are returned with a warning.
}
\examples{
# 10 nearest neighbours of Egyptian Arabic:
nearest_geo_neighbours("aeg", k=10)

# 10 nearest neighbours of Egyptian Arabic among those languages for
# which feature 13A is attested:
nearest_geo_neighbours("aeg", k=10, data=filter_WALS(feature_ID=="13A"))

# 10 nearest neighbours of Egyptian Arabic among those languages for
# which feature 13A is attested, among the Afro-Asiatic family:
nearest_geo_neighbours("aeg", k=10,
  data=filter_WALS(feature_ID=="13A" & family=="Afro-Asiatic"))
}
