% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{WALS}
\alias{WALS}
\title{World Atlas of Language Structures}
\format{A data frame with 76465 rows and 20 variables:
\describe{
\item{\code{combined_ID}}{}
\item{\code{language_ID}}{}
\item{\code{feature_ID}}{}
\item{\code{value_ID}}{}
\item{\code{code_ID}}{}
\item{\code{language}}{}
\item{\code{feature}}{}
\item{\code{value}}{}
\item{\code{iso_code}}{}
\item{\code{glottocode}}{}
\item{\code{latitude}}{}
\item{\code{longitude}}{}
\item{\code{genus}}{}
\item{\code{family}}{}
\item{\code{macroarea}}{}
\item{\code{countrycodes}}{}
\item{\code{in_WALS_100_sample}}{}
\item{\code{in_WALS_200_sample}}{}
\item{\code{source}}{}
\item{\code{contributors}}{}
}}
\source{
\url{http://wals.info}
}
\description{
The full WALS dataset.
}
